+++ packages/utils/src/index.ts
@@ -4,3 +4,4 @@ export * from './throttle';
 export * from './environment';
 export * from './focus-trap';
 export * from './keycode';
+export * from './clamp';