+++ packages/drawer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @agencecinq/drawer
 
+## 5.0.0
+
+### Minor Changes
+
+- 3594668: Spinbutton package
+
+### Patch Changes
+
+- Updated dependencies [3594668]
+  - @agencecinq/utils@5.0.0
+
 ## 4.1.3
 
 ### Patch Changes