+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @agencecinq/modal
 
+## 2.0.0
+
+### Minor Changes
+
+- 3594668: Spinbutton package
+
+### Patch Changes
+
+- Updated dependencies [3594668]
+  - @agencecinq/utils@5.0.0
+
 ## 1.1.0
 
 ### Minor Changes